+++ tests/ecc.rs
@@ -306,7 +306,7 @@ fn mul_point_works() {
 }
 
 #[test]
-fn assert_equal_point_different_y_fails() {
+fn assert_equal_point_works() {
     let rng = &mut StdRng::seed_from_u64(8349u64);
 
     let n = 1 << 4;
@@ -318,17 +318,17 @@ fn assert_equal_point_different_y_fails() {
         p2: JubJubAffine,
     }
 
+    impl DummyCircuit {
+        pub fn new(p1: JubJubAffine, p2: JubJubAffine) -> Self {
+            Self { p1, p2 }
+        }
+    }
+
     impl Default for DummyCircuit {
         fn default() -> Self {
             Self {
-                p1: JubJubAffine::from_raw_unchecked(
-                    BlsScalar::one(),
-                    BlsScalar::one(),
-                ),
-                p2: JubJubAffine::from_raw_unchecked(
-                    BlsScalar::one(),
-                    BlsScalar::zero(),
-                ),
+                p1: dusk_jubjub::GENERATOR,
+                p2: dusk_jubjub::GENERATOR,
             }
         }
     }
@@ -338,21 +338,95 @@ fn assert_equal_point_different_y_fails() {
         where
             C: Composer,
         {
-            // test: p1 != p2
             let w_p1 = composer.append_point(self.p1);
             let w_p2 = composer.append_point(self.p2);
             composer.assert_equal_point(w_p1, w_p2);
+
             Ok(())
         }
     }
 
-    let (prover, _verifier) = Compiler::compile::<DummyCircuit>(&pp, label)
+    let (prover, verifier) = Compiler::compile::<DummyCircuit>(&pp, label)
         .expect("failed to compile circuit");
 
-    let proving_result = prover.prove(rng, &Default::default());
+    // Test default works:
+    // GENERATOR = GENERATOR
+    {
+        let (proof, public_inputs) = prover
+            .prove(rng, &Default::default())
+            .expect("prover shouldn't fail");
 
-    assert!(
-        proving_result.is_err(),
-        "proving should fail because the y coordinates don't match"
-    );
+        assert_eq!(public_inputs.len(), 0);
+
+        verifier
+            .verify(&proof, &public_inputs)
+            .expect("Default circuit verification should pass");
+    }
+
+    // Test sanity:
+    // 42 * GENERATOR = 42 * GENERATOR
+    {
+        let scalar = JubJubScalar::from(42u64);
+        let p1 = dusk_jubjub::GENERATOR_EXTENDED * &scalar;
+        let p2 = dusk_jubjub::GENERATOR_EXTENDED * &scalar;
+        let circuit = DummyCircuit::new(p1.into(), p2.into());
+
+        let (proof, public_inputs) =
+            prover.prove(rng, &circuit).expect("prover shouldn't fail");
+
+        assert_eq!(public_inputs.len(), 0);
+
+        verifier
+            .verify(&proof, &public_inputs)
+            .expect("Circuit verification with equal points should pass");
+    }
+
+    // Test:
+    // GENERATOR != 42 * GENERATOR
+    {
+        let scalar = JubJubScalar::from(42u64);
+        let p1 = dusk_jubjub::GENERATOR;
+        let p2 = dusk_jubjub::GENERATOR_EXTENDED * &scalar;
+        let circuit = DummyCircuit::new(p1, p2.into());
+
+        prover
+            .prove(rng, &circuit)
+            .expect_err("prover should fail because the points are not equal");
+    }
+
+    // Test:
+    // assertion of points with different x-coordinates fails
+    {
+        let p1 = JubJubAffine::from_raw_unchecked(
+            BlsScalar::one(),
+            BlsScalar::one(),
+        );
+        let p2 = JubJubAffine::from_raw_unchecked(
+            BlsScalar::zero(),
+            BlsScalar::one(),
+        );
+        let circuit = DummyCircuit::new(p1, p2);
+
+        prover
+            .prove(rng, &circuit)
+            .expect_err("prover should fail because the x-coordinates of the points are not equal");
+    }
+
+    // Test:
+    // assertion of points with different y-coordinates fails
+    {
+        let p1 = JubJubAffine::from_raw_unchecked(
+            BlsScalar::one(),
+            BlsScalar::one(),
+        );
+        let p2 = JubJubAffine::from_raw_unchecked(
+            BlsScalar::one(),
+            BlsScalar::zero(),
+        );
+        let circuit = DummyCircuit::new(p1, p2);
+
+        prover
+            .prove(rng, &circuit)
+            .expect_err("prover should fail because the y-coordinates of the points are not equal");
+    }
 }